backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.19.0
openai==2.5.0
packaging==25.0
pillow==11.3.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stripe==13.0.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.6
